pyproject.toml
[project]
name = "midnightcommits"
version = "0.1.0"
description = "Real-time GitHub commit visualization"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiokafka>=0.10.0",
    "astral>=3.2",
    "fastapi>=0.109.0",
    "geonamescache>=3.0.0",
    "geopy>=2.4.1",
    "kafka-python>=2.0.2",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "uvicorn>=0.27.0",
    "websockets>=12.0",
]
